+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changes in 0.99.1
 
+- Add `mtbls_sync()` to synchronize a `MsBackendMetaboLights` object.
 - Add `mtbls_sync_data_files()` function to cache selected files and
   `mtbls_cached_data_files()` to list all locally cached data files.